pyproject.toml
[project]

name = "magic_cheatsheet"
version = "1.0.0"
description = " A fast and easy to use in-terminal cheatsheet generator "
authors = [
    { name = "NoLilypad" }
]
readme = "README.md"
requires-python = ">=3.8"


[project.optional-dependencies]
all = [
    "platformdirs",
    "pyyaml"
]

[project.scripts]
magic_cheatsheet = "magic_cheatsheet.cli:main"
